* Rename JsonForms to JsonFormsInit * Rename DispatchRenderer to JsonForms * Rename file of JsonFormsElement * Add clean script to test package
